LPS-86636 Change return type to int

+++ modules/apps/segments/segments-api/src/main/java/com/liferay/segments/odata/retriever/ODataRetriever.java
@@ -30,7 +30,7 @@ public interface ODataRetriever<T extends BaseModel<?>> {
 			int end)
 		throws PortalException;
 
-	public long getResultsCount(
+	public int getResultsCount(
 			long companyId, String filterString, Locale locale)
 		throws PortalException;
